Move more api logic into the api project
